# api/todo-list/server/v1/server.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "todo-list/server/v1/server.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "Todo"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/todo/{id}": {
      "get": {
        "operationId": "Todo_Get",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "Todo"
        ]
      }
    },
    "/v1/todos": {
      "get": {
        "operationId": "Todo_Index",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1IndexResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          }
        ],
        "tags": [
          "Todo"
        ]
      },
      "post": {
        "operationId": "Todo_Create",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "type": "object",
              "properties": {}
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1CreateRequest"
            }
          }
        ],
        "tags": [
          "Todo"
        ]
      },
      "put": {
        "operationId": "Todo_Update",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "type": "object",
              "properties": {}
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1UpdateRequest"
            }
          }
        ],
        "tags": [
          "Todo"
        ]
      }
    },
    "/v1/todos/{id}": {
      "delete": {
        "operationId": "Todo_Delete",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "type": "object",
              "properties": {}
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "Todo"
        ]
      }
    }
  },
  "definitions": {
    "v1CreateRequest": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string"
        },
        "desc": {
          "type": "string"
        }
      }
    },
    "v1GetResponse": {
      "type": "object",
      "properties": {
        "todo": {
          "$ref": "#/definitions/v1Todo"
        }
      }
    },
    "v1IndexResponse": {
      "type": "object",
      "properties": {
        "todos": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1Todo"
          }
        }
      }
    },
    "v1Todo": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "desc": {
          "type": "string"
        },
        "createdAt": {
          "type": "string"
        },
        "updatedAt": {
          "type": "string"
        }
      }
    },
    "v1UpdateRequest": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "desc": {
          "type": "string"
        }
      }
    }
  }
}